move and rename app selector to sapio plugin picker
