+++ src/UX/ContractCreator/SapioPluginPicker/Menu.js
@@ -1,14 +1,14 @@
 import React from 'react';
-import { AppSelector } from './AppSelector';
+import { PluginSelector } from './PluginSelector';
 export class Menu extends React.Component {
     render() {
         return (
-            <AppSelector
+            <PluginSelector
                 applications={this.props.args}
                 compiler={this.props.compiler}
                 hide={this.props.hide}
                 load_new_model={this.props.load_new_model}
-            ></AppSelector>
+            ></PluginSelector>
         );
     }
 }
